Loading bar

+++ include/tuimaker/Window.hpp
@@ -33,10 +33,10 @@ namespace tmk
 
     struct WindowSize
     {
-        int x;
-        int y;
-        int width;
-        int height;
+        size_t x;
+        size_t y;
+        size_t width;
+        size_t height;
     };
 
     class Window
